:lipstick: splits server params to individual files
